Fixed caching bug (+ moved file to correct location)
The cache now pulls the EIG value from the unique_programs list, not the acceptable_programs list: the former is a subset (not necessarily a proper subset) of the latter, so sometimes the EIG values pulled were getting pulled from the wrong list and therefore cached wrongly. 

Please Slack me for more details, I don't think this is a very clear description of the bug.